Added Joi verification and edited adding and substracting form balance

+++ src/services/transactionService.js
@@ -41,7 +41,11 @@ const getExpensesTransactionsByUser = async (userId) => {
 const deleteTransaction = async (transactionId, userId) => {
   const transaction = await Transaction.findOne({ _id: transactionId });
   await Transaction.findOneAndDelete({ _id: transactionId });
-  await updateBalance(userId, parseInt(-transaction.amount), transaction.type);
+  await updateBalance(
+    userId,
+    parseFloat(-transaction.amount),
+    transaction.type
+  );
 };
 
 const updateBalance = async (userId, amount, type) => {
@@ -49,9 +53,9 @@ const updateBalance = async (userId, amount, type) => {
   const previousUserBalance = user.balance;
   let currentUserBalance;
   if (type === "income") {
-    currentUserBalance = parseInt(previousUserBalance) + parseInt(amount);
+    currentUserBalance = parseFloat(previousUserBalance) + parseFloat(amount);
   } else {
-    currentUserBalance = parseInt(previousUserBalance) - parseInt(amount);
+    currentUserBalance = parseFloat(previousUserBalance) - parseFloat(amount);
   }
   await User.findOneAndUpdate(
     { _id: userId },
